Automate Infrastructure (#13)
I want to be able to deploy DashFloat in a way that we can control. With this in mind, Fly and other similar platforms are out of the question.

I'm choosing to deploy to GCP for the time being but I've decided to use Pulumi to automate the infrastructure in order to maintain the option of moving to other cloud platforms.